+++ src/Tag/Input/Color.php
@@ -6,15 +6,13 @@ namespace Yiisoft\Html\Tag\Input;
 
 use Yiisoft\Html\Tag\Base\InputTag;
 
+/**
+ * @link https://html.spec.whatwg.org/multipage/input.html#color-state-(type=color)
+ */
 final class Color extends InputTag
 {
     protected function prepareAttributes(): void
     {
-        $this->attributes['type'] = $this->getType();
-    }
-
-    protected function getType(): string
-    {
-        return 'color';
+        $this->attributes['type'] = 'color';
     }
 }